Encapsulate the bit arrangement of tilemap entries
A class `Tilemap` is used to compile the data for a tilemap entry into the binary 5 byte data. As just one benefit, this allows the code to do direct arithmetic on the x/y values instead of handling value and the negative flag in parallel.

The "pal" key in the json data has changed to a "r" key which holds an array on the form [palette index, priority] which correspond to the bit fields in the tilemap dword, where the palette index is a 3 bit wide field, and priority is a 2 bit wide field.