Add safeguards for fairy pond boomerang/shield/bow upgrades
The items that trigger the upgrades are changed such that:
- Red Boomerang trigger in order to give back Red Boomerang
- Red Shield trigger in order to give back Red Shield
- Bow with Silver arrows trigger in order to give back Bow with Silver arrows

The upgraded item were chosen over having the initial item (Blue Boomerang, Blue Shield, Bow with normal arrows) give itself because the dialog text speaks of the upgraded item. No change were made to the sword upgrade since the code for this "hides" deeper.