Support autotracking total heart pieces collected
ALTTP stats stored in $420 to $466 in ALTTP SRAM contains among other things
a total heart piece counter.  This is incremented in ALTTP when picking up
a heart piece already, and this is used on the file select screen in Z3R.
However in SMZ3 only the item addresses $300 to $3FF are copied to a buffer
while in SM.

This change adds an extra buffer for ALTTP stats and copies these addresses
to and from the buffer in the existing transition code, as well as adding
the same logic when collecting a heart piece in SM to increment this global
counter.  This allows the number of collected heart pieces and heart containers
to be autotracked effectively, as well as making the other stats available
in SM for any need.