Chest key counters for autotracking
This is a port of https://github.com/KatDevsGames/z3randomizer/pull/90 which adds
support for a new stats cache of chest small key counts for ALTTP dungeons.
This allows for much easier autotracking of dungeon items found in non-keysanity
mode, and found small key counts in keysanity mode.  We add support for the SM
side as well to increment these stat counters when collecting a small key in SM
in keysanity mode.