Clear mw seed data and re-init on new game
This change fixes an issue when deleting your save game data after receiving a bunch of MW items.
Without this, the game thinks it already received all the sent items and they don't get re-sent
if you delete your save and start a new one.  This clears the MW seed data and calls the init
subroutine which will see that the seed data doesn't match and re-initialize everything.