Fix autotracking dungeon chest counters in multiworld
This fixes an issue where the dungeon chest counters used for autotracking in
non-keysanity mode were not being incremented when you found an item for another
player in multiworld.  They were also incorrectly being incremented when you
received an item from another player if you were inside a dungeon at the time.
This moves the chest counter part to a subroutine and calls it when sending an
item to another player, and skips it in the item pickup part if we're receiving
an item from another player during multiworld.